test(testing): SourceManagement tests вынесены в отдельные suite files
service-level, app-flow, screen state и screen controller tests для source management выделены из RSSReaderTests.swift в самостоятельные feature-oriented suites
