+++ README.md
@@ -332,6 +332,7 @@
 - [x] вынести `Shell` и `WebViewScreen` navigation/state tests из `RSSReaderTests.swift` в отдельные suite files, чтобы shell-level routing и embedded browser flow были отделены от `ArticlesScreen` и app settings/persistence блоков;
 - [x] вынести `ArticlesScreen` state/controller/mutation/presentation tests из `RSSReaderTests.swift` в отдельные suite files, чтобы list flow был разложен по screen state, controller orchestration, mutation reducer и presentation helpers;
 - [x] вынести `Settings`/`AppSettings` persistence/service/presentation/state/controller/theme tests из `RSSReaderTests.swift` в отдельные suite files, чтобы app settings flow был разложен по repository/service, screen presentation/state, controller orchestration и theme policy;
+- [x] вынести `SourceManagement` service/app-flow/state/controller tests из `RSSReaderTests.swift` в отдельные suite files, чтобы source onboarding и source editing flow были разложены по service-layer, app-level launch context, screen state и screen controller orchestration;
 - [ ] зафиксировать целевой sync scope: в CloudKit должны синхронизироваться `Feed`, `Folder`, `ArticleState` и `AppSettings`, а `Article` и `FeedFetchLog` должны оставаться только локальным storage слоем;
 - [ ] провести audit CloudKit compatibility для `AppSettings`, `Feed` и `Folder`: проверить `@Attribute(.unique)`, relationship semantics и текущие invariants, которые могут блокировать SwiftData sync;
 - [ ] провести audit CloudKit compatibility для `ArticleState`, `Article` и `FeedFetchLog`: проверить `#Unique`, nonoptional relationships, delete rules и закрепить границу между sync-backed сущностями и локальным article cache;
