+++ RSSReaderTests/SourceManagement/SourceManagementAppFlowTests.swift
@@ -1,0 +1,61 @@
+import Foundation
+import Testing
+@testable import RSSReader
+
+@Suite("Source Management / App Flow")
+@MainActor
+struct SourceManagementAppFlowTests {
+    @Test
+    func sourceManagementPresentationStateUsesSeparateModalFlowAndDoesNotResetReadingShellContext() throws {
+        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
+        let appState = AppState()
+        let feedID = UUID()
+        let articleID = UUID()
+
+        harness.dependencies.showFeed(id: feedID, using: appState)
+        harness.dependencies.selectArticle(id: articleID, using: appState)
+
+        harness.dependencies.showSourceManagement(using: appState)
+
+        #expect(appState.isPresentingSourceManagementScreen)
+        #expect(appState.isPresentingSettingsScreen == false)
+        #expect(appState.selectedSidebarSelection == .feed(feedID))
+        #expect(appState.selectedArticleID == articleID)
+        #expect(appState.selectedDetailRoute == .article(articleID))
+
+        harness.dependencies.dismissSourceManagement(using: appState)
+
+        #expect(appState.isPresentingSourceManagementScreen == false)
+        #expect(appState.isPresentingSettingsScreen == false)
+        #expect(appState.sourceManagementLaunchContext == .entry)
+        #expect(appState.selectedSidebarSelection == .feed(feedID))
+        #expect(appState.selectedArticleID == articleID)
+        #expect(appState.selectedDetailRoute == .article(articleID))
+    }
+
+    @Test
+    func sourceManagementPresentationStateTracksSidebarEditLaunchContextWithoutResettingReadingShellContext() throws {
+        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
+        let appState = AppState()
+        let feedID = UUID()
+        let articleID = UUID()
+
+        harness.dependencies.showFeed(id: feedID, using: appState)
+        harness.dependencies.selectArticle(id: articleID, using: appState)
+        harness.dependencies.showFeedEditor(id: feedID, using: appState)
+
+        #expect(appState.isPresentingSourceManagementScreen)
+        #expect(appState.sourceManagementLaunchContext == .editFeed(feedID))
+        #expect(appState.selectedSidebarSelection == .feed(feedID))
+        #expect(appState.selectedArticleID == articleID)
+        #expect(appState.selectedDetailRoute == .article(articleID))
+
+        harness.dependencies.dismissSourceManagement(using: appState)
+
+        #expect(appState.isPresentingSourceManagementScreen == false)
+        #expect(appState.sourceManagementLaunchContext == .entry)
+        #expect(appState.selectedSidebarSelection == .feed(feedID))
+        #expect(appState.selectedArticleID == articleID)
+        #expect(appState.selectedDetailRoute == .article(articleID))
+    }
+}
